+++ dev/examples/Animation-animate.tsx
@@ -1,6 +1,7 @@
 import * as React from "react"
 import { useEffect, useState } from "react"
 import { motion, motionValue, useAnimate } from "framer-motion"
+import { frame } from "framer-motion"
 
 /**
  * An example of the tween transition type