+++ packages/framer-motion/src/animation/optimized-appear/types.ts
@@ -1,5 +1,18 @@
-import { Batcher } from "../../frameloop/types"
-import { MotionValue } from "../../value"
+import type { Batcher } from "../../frameloop/types"
+import type { MotionValue } from "../../value"
+
+export type HandoffFunction = (
+    storeId: string,
+    valueName: string,
+    /**
+     * Legacy arguments. This function is inlined as part of SSG so it can be there's
+     * a version mismatch between the main included Motion and the inlined script.
+     *
+     * Remove in early 2024.
+     */
+    _value: MotionValue,
+    _frame: Batcher
+) => null | number
 
 /**
  * The window global object acts as a bridge between our inline script
@@ -7,11 +20,7 @@ import { MotionValue } from "../../value"
  */
 declare global {
     interface Window {
-        HandoffAppearAnimations?: (
-            storeId: string,
-            valueName: string,
-            value: MotionValue,
-            sync: Batcher
-        ) => number
+        HandoffAppearAnimations?: HandoffFunction
+        HandoffComplete?: boolean
     }
 }